requirements.txt
gradio
numpy
openai
pydantic
bardapi==0.1.23a
urllib3
python-dotenv